feat: 활동 내역 컴포넌트 api 연결
